+++ client/src/pages/Reg.tsx
@@ -7,24 +7,39 @@ import { Link } from "react-router-dom";
 import cookie from "../utils/cookie";
 
 function Login() {
-  const [formData, setFormData] = useState({ email: "", password: "" });
+  const [formData, setFormData] = useState({
+    email: "",
+    password: "",
+    password2: "",
+    username: "",
+  });
+
   const [error, setError] = useState<string | null>(null);
 
   const formDataSchema = Joi.object({
     email: Joi.string().email({ tlds: false }),
     password: Joi.string().min(5),
+    password2: Joi.string().min(5).valid(Joi.ref("password")),
+    username: Joi.string().alphanum().min(3),
   });
 
-  const submitHandler = () => {
+  const submitHandler = async () => {
     const { error } = formDataSchema.validate(formData, { abortEarly: false });
     if (error) return setError(error.message);
 
-    userService.userLogin(formData.email, formData.password).then((data) => {
-      cookie.clear();
+    const resData = await userService.userReg(
+      formData.username,
+      formData.email,
+      formData.password
+    );
 
-      if (data.message) return console.error(data.message);
-      if (data.token) cookie.setData({ token: data.token });
-    });
+    if (resData.message) {
+      console.error(resData.message);
+    }
+
+    if (resData.token) {
+      cookie.setData({ token: resData.token });
+    }
   };
 
   return (
@@ -43,12 +58,20 @@ function Login() {
           setError={setError}
         />
       )}
-      <UI.Modal name="login" className="active">
+      <UI.Modal name="register" className="active">
         <UI.Input
           placeholder="email"
           type="email"
           value={formData.email}
           onChange={(e) => setFormData({ ...formData, email: e.target.value })}
+        />
+        <UI.Input
+          placeholder="name"
+          type="text"
+          value={formData.username}
+          onChange={(e) =>
+            setFormData({ ...formData, username: e.target.value })
+          }
         />
         <UI.Input
           placeholder="password"
@@ -58,8 +81,16 @@ function Login() {
             setFormData({ ...formData, password: e.target.value })
           }
         />
+        <UI.Input
+          placeholder="password"
+          type="text"
+          value={formData.password2}
+          onChange={(e) =>
+            setFormData({ ...formData, password2: e.target.value })
+          }
+        />
         <span>
-          or <Link to="/reg">register</Link>
+          or <Link to="/login">login</Link>
         </span>
         <UI.Button onClick={submitHandler}>enter</UI.Button>
       </UI.Modal>