+++ client/src/router/main.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import Home from "../pages/Home";
 import Login from "../pages/Login";
+import Register from "../pages/Reg";
 
 import { Navigate } from "react-router-dom";
 
@@ -13,6 +14,7 @@ export const privateRoutes: RouteT[] = [
 export const publicRoutes: RouteT[] = [
   { path: "/", component: <Home /> },
   { path: "/login", component: <Login /> },
+  { path: "/reg", component: <Register /> },
   // { path: "*", component: <Navigate to="/login" /> },
 ];