+++ run3205.sh
@@ -1,1 +1,0 @@
-{ cd server && npm run build & cls & npm run start; } & { cd client && npm run start } &