+++ server/src/errors/API.error.ts
@@ -7,7 +7,7 @@ export class ApiError extends Error {
   }
 
   static badRequest(message: string) {
-    return new ApiError(404, message);
+    return new ApiError(400, message);
   }
   static internal(message: string) {
     return new ApiError(500, message);